added music info like artist, song name, album from spotify web api response object
